+++ Automate/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed machine scan failing if another mod adds broken furniture or terrain features.
+
 ## 1.27.2
 Released 18 June 2022 for SMAPI 3.14.0 or later.